cv-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
weasyprint>=63.0
jinja2>=3.1.0
python-docx>=1.1.0
pydantic>=2.10.0
httpx>=0.28.0
pillow>=11.0.0
python-multipart>=0.0.18
python-jose[cryptography]>=3.3.0
